feat: introduce 'messagesonhead.toggle' permission + move 'MessageStackRepository' tp 'services' package
