+++ src/main/java/mrquackduck/messagesonhead/configuration/Permissions.java
@@ -2,5 +2,6 @@ package mrquackduck.messagesonhead.configuration;
 
 public class Permissions {
     public static final String SHOW = "messagesonhead.show";
+    public static final String TOGGLE = "messagesonhead.toggle";
     public static final String ADMIN = "messagesonhead.admin";
 }
