+++ src/main/resources/plugin.yml
@@ -13,6 +13,9 @@ permissions:
   messagesonhead.show:
     description: Whether messages will appear above player's head
     default: true
+  messagesonhead.toggle:
+    description: Allows to use the '/moh toggle' command
+    default: true
   messagesonhead.admin:
     description: Allows to use staff commands of the plugin
     default: op
